refactor: move DataVersion out of ondisk into its own module
# Summary

`DataVersion`, `VersionInfo` and the `DATA_VERSION` constant now live in a
top-level `data_version` module instead of inside `ondisk`, breaking the
dependency cycle between `config` and `ondisk`.

# Details

`config` needed `DATA_VERSION` to derive the on-disk directory layout, while
`ondisk` needed `RaftConfig` to locate the data directory. That mutual
dependency meant neither module could be reasoned about, or later extracted,
in isolation.

The version types are pure data with no sled dependency, so they sit below
both modules. `Header` stays in `ondisk` because it is the sled-serialized
record of the version, not the version itself. The remaining edges are now
one-way: `config` and `ondisk` both depend on `data_version`, and `ondisk`
depends on `config`.

This also drops the incidental `sm_v003` and `leveled_store` edges into
`ondisk`, which existed only to reach `DATA_VERSION`.

`DataVersion` and `DATA_VERSION` move from `databend_meta_raft_store::ondisk`
to `databend_meta_raft_store::data_version`; call sites in the service crate
are updated accordingly.
