+++ crates/server/raft-store/src/lib.rs
@@ -44,6 +44,7 @@
 
 pub mod applier;
 pub mod config;
+pub mod data_version;
 pub mod immutable_compactor;
 pub mod key_spaces;
 pub mod leveled_store;
