+++ crates/server/raft-store/tests/it/config.rs
@@ -15,7 +15,7 @@
 use databend_meta_raft_store::MetaStartupError;
 use databend_meta_raft_store::config::RaftConfig;
 use databend_meta_raft_store::config::get_default_raft_advertise_host;
-use databend_meta_raft_store::ondisk::DATA_VERSION;
+use databend_meta_raft_store::data_version::DATA_VERSION;
 use databend_meta_runtime_api::TokioRuntime;
 use databend_meta_types::Endpoint;
 use pretty_assertions::assert_eq;
