+++ crates/server/service/src/meta_node/meta_node.rs
@@ -27,7 +27,7 @@ use databend_meta_client::RequestFor;
 use databend_meta_raft_store::MetaStartupError;
 use databend_meta_raft_store::StateMachineFeature;
 use databend_meta_raft_store::config::RaftConfig;
-use databend_meta_raft_store::ondisk::DATA_VERSION;
+use databend_meta_raft_store::data_version::DATA_VERSION;
 use databend_meta_raft_store::raft_log_v004::RaftLogStat;
 use databend_meta_raft_store::utils::seq_marked_to_seqv;
 use databend_meta_runtime_api::JoinHandle;
