+++ crates/server/service/src/meta_node/meta_node_status.rs
@@ -14,7 +14,7 @@
 
 use std::collections::BTreeMap;
 
-use databend_meta_raft_store::ondisk::DataVersion;
+use databend_meta_raft_store::data_version::DataVersion;
 use databend_meta_raft_store::raft_log_v004::RaftLogStat;
 use databend_meta_types::node::Node;
 use databend_meta_types::raft_types::LogId;
